Non-repeating update
Changed code to load lists on first load and only refresh when they get empty instead of on every button click. Reduces frequency of reading the csv file and eliminates duplicates. (Technically, there could be a duplicate if it's the last item and is randomly chosen as the first on refresh)